# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SaaS Bonus System API
  description: |
    API для управления бонусными программами с поддержкой мультитенантности и интеграцией с Telegram ботами.
    
    ## Основные возможности:
    - Управление проектами (тенантами)
    - Управление пользователями и бонусами
    - Webhook интеграция для внешних систем
    - Управление Telegram ботами
    - Аналитика и статистика
    - Реферальная программа
    - Многоуровневая система бонусов
  version: 1.0.0
  contact:
    email: support@saas-bonus.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.saas-bonus.com
    description: Production server
  - url: http://localhost:5006
    description: Development server
security:
  - ClerkAuth: []
...
paths:
  /api/webhook/{webhookSecret}:
    post:
      tags:
        - Webhook
      summary: Обработка webhook запросов
      description: |
        Универсальный endpoint для обработки различных действий:
        - Регистрация пользователей
        - Начисление бонусов за покупки
        - Списание бонусов
        - Интеграция с Tilda
      parameters:
        - name: webhookSecret
          in: path
          required: true
          schema:
            type: string
          description: Уникальный секретный ключ проекта
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/RegisterUserRequest'
                - $ref: '#/components/schemas/PurchaseRequest'
                - $ref: '#/components/schemas/SpendBonusesRequest'
                - $ref: '#/components/schemas/TildaOrderRequest'
      responses:
        '200':
          description: Успешная обработка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'
        '201':
          description: Пользователь создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCreatedResponse'
        '400':
          description: Некорректные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неверный webhook secret
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Проект деактивирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Превышен лимит запросов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    ClerkAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk JWT токен для аутентификации
  schemas:
...
    RegisterUserRequest:
      type: object
      required:
        - action
        - payload
      properties:
        action:
          type: string
          enum: [register_user]
        payload:
          type: object
          properties:
            email:
              type: string
              format: email
            phone:
              type: string
            firstName:
              type: string
            lastName:
              type: string
            birthDate:
              type: string
              format: date
            utmSource:
              type: string
            utmMedium:
              type: string
            utmCampaign:
              type: string
            utmContent:
              type: string
            utmTerm:
              type: string
            referralCode:
              type: string
    PurchaseRequest:
      type: object
      required:
        - action
        - payload
      properties:
        action:
          type: string
          enum: [purchase]
        payload:
          type: object
          required:
            - amount
            - orderId
          properties:
            email:
              type: string
              format: email
            phone:
              type: string
            amount:
              type: number
              minimum: 0
            orderId:
              type: string
            description:
              type: string
    SpendBonusesRequest:
      type: object
      required:
        - action
        - payload
      properties:
        action:
          type: string
          enum: [spend_bonuses]
        payload:
          type: object
          required:
            - amount
            - orderId
          properties:
            email:
              type: string
              format: email
            phone:
              type: string
            amount:
              type: number
              minimum: 0
            orderId:
              type: string
            description:
              type: string
    TildaOrderRequest:
      type: array
      items:
        type: object
        properties:
          name:
            type: string
          email:
            type: string
            format: email
          phone:
            type: string
          payment:
            type: object
            properties:
              amount:
                type: string
              orderid:
                type: string
              systranid:
                type: string
              products:
                type: array
                items:
                  type: object
                  properties:
                    name:
                      type: string
                    quantity:
                      type: number
                    price:
                      type: number

    # Models
...
    User:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
          format: date
        telegramId:
          type: string
        telegramUsername:
          type: string
        isActive:
          type: boolean
        totalPurchases:
          type: number
        currentLevel:
          type: string
        referralCode:
          type: string
        referredBy:
          type: string
        registeredAt:
          type: string
          format: date-time
...
    WebhookResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
    UserCreatedResponse:
      type: object
      properties:
        success:
          type: boolean
        user:
          $ref: '#/components/schemas/User'
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          default: false
        error:
          type: string
        code:
          type: string
        details:
          type: object
...